pyproject.toml
[build-system]
requires = [ "setuptools>=61.0",]
build-backend = "setuptools.build_meta"

[project]
name = "airtrain"
version = "0.1.61"
description = "A platform for building and deploying AI agents with structured skills"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [ "Development Status :: 3 - Alpha", "Intended Audience :: Developers", "License :: OSI Approved :: MIT License", "Operating System :: OS Independent", "Programming Language :: Python :: 3", "Programming Language :: Python :: 3.8", "Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10",]
dependencies = [ "pydantic>=2.10.6", "openai>=1.60.1", "python-dotenv>=1.0.1", "PyYAML>=6.0.2", "firebase-admin>=6.6.0", "loguru>=0.7.3", "requests>=2.32.3", "boto3>=1.36.6", "together>=1.3.13", "anthropic>=0.45.0", "groq>=0.15.0", "cerebras-cloud-sdk>=1.19.0", "google-genai>=1.0.0", "fireworks-ai>=0.15.12", "google-generativeai>=0.8.4", "click>=8.0.0", "rich>=13.3.1", "prompt-toolkit>=3.0.36", "colorama>=0.4.6", "typer>=0.9.0", "posthog>=3.7.0",]
[[project.authors]]
name = "Dheeraj Pai"
email = "helloworldcmu@gmail.com"

[project.optional-dependencies]
dev = [ "black>=24.10.0", "flake8>=7.1.1", "isort>=5.13.0", "mypy>=1.9.0", "pytest>=7.0.0", "twine>=4.0.0", "build>=0.10.0", "types-PyYAML>=6.0", "types-requests>=2.31.0", "types-Markdown>=3.5.0", "toml>=0.10.2",]

[project.urls]
Homepage = "https://github.com/rosaboyle/airtrain.dev"
Documentation = "https://docs.airtrain.dev/"

[project.scripts]
airtrain = "airtrain.cli.main:main"

[tool.black]
line-length = 88
target-version = [ "py38",]
include = "\\.pyi?$"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
verbose = true
showlocals = true
log_cli = true
log_cli_level = "INFO"
addopts = "--tb=native"
testpaths = [ "tests",]
filterwarnings = [ "ignore::DeprecationWarning", "ignore::PendingDeprecationWarning", "default::UserWarning",]
